chore: remove legacy evaluator
