+++ internal/config/testdata/default.yml
@@ -2,19 +2,9 @@
 #   level: INFO
 #   grpc_level: ERROR
 
-# ui:
-#   enabled: true
-
 # cors:
 #   enabled: false
 #   allowed_origins: "*"
-
-# cache:
-#   enabled: false
-#   backend: memory
-#   ttl: 60s
-#   memory:
-#     eviction_interval: 5m # Evict Expired Items Every 5m
 
 # server:
 #   protocol: http
@@ -22,6 +12,3 @@
 #   https_port: 443
 #   http_port: 8080
 #   grpc_port: 9000
-
-# db:
-#   url: file:/var/opt/flipt/flipt.db
