+++ internal/server/authn/method/token/server.go
@@ -49,6 +49,8 @@ func (s *Server) AllowsNamespaceScopedAuthentication(ctx context.Context) bool {
 // Given the token is created successfully, the generate clientToken string is returned.
 // Along with the created Authentication, which includes it's identifier and associated timestamps.
 func (s *Server) CreateToken(ctx context.Context, req *auth.CreateTokenRequest) (*auth.CreateTokenResponse, error) {
+	s.logger.Debug("CreateToken", zap.Any("request", req))
+
 	metadata := req.Metadata
 	if metadata == nil {
 		metadata = map[string]string{}
