+++ internal/server/authn/server.go
@@ -122,7 +122,7 @@ func (s *Server) AllowsNamespaceScopedAuthentication(ctx context.Context) bool {
 // GetAuthenticationSelf returns the Authentication which was derived from the request context.
 func (s *Server) GetAuthenticationSelf(ctx context.Context, _ *emptypb.Empty) (*auth.Authentication, error) {
 	if auth := authmiddlewaregrpc.GetAuthenticationFrom(ctx); auth != nil {
-		s.logger.Debug("GetAuthentication", zap.String("id", auth.Id))
+		s.logger.Debug("GetAuthenticationSelf", zap.String("id", auth.Id))
 
 		return auth, nil
 	}
@@ -132,11 +132,14 @@ func (s *Server) GetAuthenticationSelf(ctx context.Context, _ *emptypb.Empty) (*
 
 // GetAuthentication returns the Authentication identified by the supplied id.
 func (s *Server) GetAuthentication(ctx context.Context, r *auth.GetAuthenticationRequest) (*auth.Authentication, error) {
+	s.logger.Debug("GetAuthentication", zap.String("id", r.Id))
+
 	return s.store.GetAuthenticationByID(ctx, r.Id)
 }
 
 // ListAuthentications produces a set of authentications for the provided method filter and pagination parameters.
 func (s *Server) ListAuthentications(ctx context.Context, r *auth.ListAuthenticationsRequest) (*auth.ListAuthenticationsResponse, error) {
+	s.logger.Debug("ListAuthentications", zap.Any("request", r))
 	req := &storage.ListRequest[storageauth.ListAuthenticationsPredicate]{
 		QueryParams: storage.QueryParams{
 			Limit:     uint64(r.Limit),
