+++ internal/server/evaluation/data/server.go
@@ -129,6 +129,8 @@ func (srv *Server) EvaluationSnapshotNamespace(ctx context.Context, r *evaluatio
 		ifNoneMatch  string
 	)
 
+	srv.logger.Debug("EvaluationSnapshotNamespace", zap.String("namespace", namespaceKey), zap.String("reference", reference))
+
 	md, ok := metadata.FromIncomingContext(ctx)
 	if ok {
 		// get If-None-Match header from request
