+++ rpc/flipt/scoped.go
@@ -24,10 +24,3 @@ func (req *DeleteNamespaceRequest) GetNamespaceKey() string {
 func (req *UpdateNamespaceRequest) GetNamespaceKey() string {
 	return req.Key
 }
-
-func (x *BatchEvaluationRequest) GetNamespaceKeys() (keys []string) {
-	for _, r := range x.Requests {
-		keys = append(keys, r.NamespaceKey)
-	}
-	return
-}
