+++ sdk/go/http/flipt.sdk.gen.go
@@ -20,70 +20,6 @@ type FliptClient struct {
 	addr   string
 }
 
-func (x *FliptClient) Evaluate(ctx context.Context, v *flipt.EvaluationRequest, _ ...grpc.CallOption) (*flipt.EvaluationResponse, error) {
-	var body io.Reader
-	var values url.Values
-	reqData, err := protojson.Marshal(v)
-	if err != nil {
-		return nil, err
-	}
-	body = bytes.NewReader(reqData)
-	req, err := http.NewRequestWithContext(ctx, http.MethodPost, x.addr+fmt.Sprintf("/api/v1/namespaces/%v/evaluate", v.NamespaceKey), body)
-	if err != nil {
-		return nil, err
-	}
-	req.URL.RawQuery = values.Encode()
-	resp, err := x.client.Do(req)
-	if err != nil {
-		return nil, err
-	}
-	defer resp.Body.Close()
-	var output flipt.EvaluationResponse
-	respData, err := io.ReadAll(resp.Body)
-	if err != nil {
-		return nil, err
-	}
-	if err := checkResponse(resp, respData); err != nil {
-		return nil, err
-	}
-	if err := (protojson.UnmarshalOptions{DiscardUnknown: true}).Unmarshal(respData, &output); err != nil {
-		return nil, err
-	}
-	return &output, nil
-}
-
-func (x *FliptClient) BatchEvaluate(ctx context.Context, v *flipt.BatchEvaluationRequest, _ ...grpc.CallOption) (*flipt.BatchEvaluationResponse, error) {
-	var body io.Reader
-	var values url.Values
-	reqData, err := protojson.Marshal(v)
-	if err != nil {
-		return nil, err
-	}
-	body = bytes.NewReader(reqData)
-	req, err := http.NewRequestWithContext(ctx, http.MethodPost, x.addr+fmt.Sprintf("/api/v1/namespaces/%v/batch-evaluate", v.NamespaceKey), body)
-	if err != nil {
-		return nil, err
-	}
-	req.URL.RawQuery = values.Encode()
-	resp, err := x.client.Do(req)
-	if err != nil {
-		return nil, err
-	}
-	defer resp.Body.Close()
-	var output flipt.BatchEvaluationResponse
-	respData, err := io.ReadAll(resp.Body)
-	if err != nil {
-		return nil, err
-	}
-	if err := checkResponse(resp, respData); err != nil {
-		return nil, err
-	}
-	if err := (protojson.UnmarshalOptions{DiscardUnknown: true}).Unmarshal(respData, &output); err != nil {
-		return nil, err
-	}
-	return &output, nil
-}
-
 func (x *FliptClient) GetNamespace(ctx context.Context, v *flipt.GetNamespaceRequest, _ ...grpc.CallOption) (*flipt.Namespace, error) {
 	var body io.Reader
 	if v.Key == "" {
